server/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
google-generativeai

#pip install -r requirements.txt
